pools ISR first pass
